Remove import cycle for CLI server
